refactor: tool-call-parserをモジュール構造に分割
729行のモノリス(tool-call-parser.ts)を責務ごとに分離:
- types.ts / utils.ts: 型定義と共通ユーティリティ
- detector.ts: runtimeInfoからのデリミタ検出（KNOWN_TOOL_PARSER_DELIMITERS一元管理）
- content-parsers.ts: ContentParserインターフェースによるプラグイン構造（json/gemma4/pythonic/xml）
- index.ts: parseToolCallsメインロジック（フォールバックチェーン維持）
- tool-formatter.ts: ツール定義のテキストフォーマッタ

Python側(token_utils.py)のKNOWN_TOOL_PARSERSテーブルを削除し、
パーサー種別→デリミタのマッピングをTS側に一元化。
MiniCPM5のXMLフォーマット対応も含む。
